+++ perspective.el
@@ -665,6 +665,11 @@ perspective's local variables are set.
 If NORECORD is non-nil, do not update the
 `persp-last-switch-time' for the switched perspective."
   (interactive "i")
+  ;; TODO: See https://github.com/nex3/perspective-el/issues/163 for the source
+  ;; of this workaround. It would be good to investigate exactly why recursive
+  ;; minibuffers break Perspective.
+  (when (> (minibuffer-depth) 0)
+    (user-error "Perspective operations in recursive minibuffers not supported"))
   (unless (persp-valid-name-p name)
     (setq name (persp-prompt (and (persp-last) (persp-name (persp-last))))))
   (if (and (persp-curr) (equal name (persp-current-name))) name